Replace embedded jquery files with Haskell libraries
To reduce the number of duplicate jquery copies shipped across Haskell
libraries, replace the embedded copies of jquery and jquery-flot with
the Haskell libraries js-jquery and js-jflot.

The only non-obvious change is the upgrade of jflot from 0.7 to 0.8,
which splits the time sub-library into a separate library, hence the
need to list it in index.html.