throw startServer exceptions synchronously
Error handling doesn't currently work with `forkServer`:
```
do
  s <- Just <$> forkServer "localhost" 8000 `catchAny` \e -> pure Nothing
```

The above code will not manage to deal with "port already in use" or similar, since the exceptions are being thrown asynchronously to the main thread.

To address that we use an Async to hold the server, so there's no longer a need to throw the exception back to the main thread.